gpu: nvgpu: posix: fix MISRA bugs in COND_WAIT
Fix MISRA 5.3 violation for hiding the variable "ret."

Fix MISRA 10.1 violation in the NVGPU_COND_WAIT() macro. The timeout
value was being used as a boolean for the ? operator. Compare to 0
instead.

Fix MISRA 14.3 violation for invariant condition.

Fix MISRA 14.4 violation for using 0 for a boolean in the while
condition.

JIRA NVGPU-3329

Change-Id: I874aa66abb8771f9855ba4312ea068603d5b2e7b
Signed-off-by: Philip Elcan <pelcan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2109471
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Alex Waterman <alexw@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>