dts: t23x: remove max-codec-slot entry from hda
This property is redundant as HDA driver already has a fallback option when
the property is not specified. Probe call, azx_probe_codecs(), loops over
the given maximum codec slots(HDA_MAX_CODECS) and registers the devices.
If needed in future, the value of HDA_MAX_CODECS can be increased as per
the available slots.

Bug 200454686

Change-Id: I0f75690493014172f4916158e7fce59465beef09
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2117632
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>