video: tegra: support base colorimetry formats
This change extends the support to switch colorimentry format
by adding base colorimetry for SMPTE170M_ITU601 and ITU709.
Earlier support was for extended colorimetry but base
colorimetry was always default if extended is not mentioned.
With this change, usermode can now change base colorimetry as
well. When not mentioned it will use default as earlier.

Bug 200352101

Change-Id: Ifa1a90012b25b38fad89b6164c5b4903f8278c5e
Signed-off-by: Prafull Suryawanshi <prafulls@nvidia.com>
Signed-off-by: Michael Frydrych <mfrydrych@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2255106
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>