video: tegra: host: Load VIC firmware from firmware root
VIC firmware for T19x/T23x has moved to the root firmware
directory from the chip-specific subdirectory. Make the
corresponding change here.

A new device data field is added to specify that the firmware
is not loaded from a subdirectory. Hopefully in the future
this will be true for all firmware and we can drop the flag.

Bug 2173351

Change-Id: I9e225caa1fca4752537fb9fe425d5486230d7c21
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2132391
GVS: Gerrit_Virtual_Submit
Tested-by: Antoine Chauveau <achauveau@nvidia.com>
Reviewed-by: Antoine Chauveau <achauveau@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>