qspi: enable clock always
Runtime PM is not a POR for Automotive Platforms. The property
"nvidia,clock-always-on" enables the clock irrespective of whether a
transaction is requested or not.

This leads to disabling of the runtime power saving which was acheieved
with disabling the clock when the clock is not needed. Ot also leads to
incresing the read/write perf a bit because read/write accesses do not need
to enable clocks which would be needed otherwise.

Change-Id: I04d27cb817e6fa3e403733cc8427e4c7c174073d
Signed-off-by: Vipin Kumar <vipink@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2135691
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ashutosh Patel <ashutoshp@nvidia.com>
Reviewed-by: Nitin Sehgal <nsehgal@nvidia.com>
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>