gpu: nvgpu: remove unreferenced typedefs
MISRA Advisory Rule 2.3 states that a project should not contain
unused type declarations.

The justification for this rule is that a code reviewer may not
know if such a declaration was left behind inadvertently.

The nvgpu deviation record for Advisory Rule 2.3 states that no such
violations will remain that fit this description.

This change removes several unreferenced typedefs so that this
requirement is met.

JIRA NVGPU-3798

Change-Id: I852b5d1fe8d6beb12e6b93219e3101d806a88a39
Signed-off-by: Scott Long <scottl@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2150415
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Thomas Fleury <tfleury@nvidia.com>
Reviewed-by: Alex Waterman <alexw@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>