pinctrl: tegra: Save and restore pinctrl state for GPIO requests
For Tegra186 and Tegra194, when a GPIO is requested, the pinmux for the
associated pin is configured as a GPIO after the direction of the GPIO
is configured. When the GPIO is freed the pinmux for the pin is
reconfigured as an SFIO regardless of whether the pin was configured as
a GPIO or SFIO before it was requested. It should be noted that Tegra210
does not behave this way and restores the state of the pin on freeing
the GPIO.

Update the Tegra pinctrl driver so that the state of the pinmux is
restored on freeing a GPIO. This is achieved by calling the function
pinctrl_gpio_save_config() when a GPIO is requested to save the pinmux
state and calling the function pinctrl_gpio_restore_config() on freeing
the GPIO to restore the pinmux state.

Bug 2414131

Change-Id: Iec8e02007cefc8c5eff57be731007595080ed8cb
Signed-off-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2155336
(cherry picked from commit 4ee73214c5fd5176be59318e8f1059f69b9e510a)
Reviewed-on: https://git-master.nvidia.com/r/2164161
(cherry picked from commit 10009737135c2500859d2e7b45b1bdcc14222073)
Reviewed-on: https://git-master.nvidia.com/r/2181965
(cherry picked from commit 571905532ce8d2cb6a9c3467ce1e1b402e766fe6)
[nicolinc: inserted before sfsel-unset code in request_enable() so
 that save_config() saves sfsel bit; then replaced sfet set code in
 disable_free() since restore_config() would restore the sfsel bit]
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2428812
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>