dts: t19x: update clock bindings for AHUB modules
Currently DT clock bindings of I2S, DMIC nodes pass parent or sync clocks.
Module drivers get clock handles for these and manage clock relations.
Instead, "assigned-clocks" and "assigned-clock-parents" clock bindings
can be used to setup clock parents. Separate entries for sync clock is not
necessary, as it is one of the source as parent for respective module
clocks. But to be backward compatible with AUTO platforms these clocks
cannot be removed. This patch uses above additional clock bindings for AHUB
modules, which will be used for L4T/Android as of now. This would help in
upstream alignment. The tegra-alt drivers are cleaned up in another patch
in the series. The "assigned-clock-rates" binding is used to initialize
the clocks with default rates.

Only for xbar, parent clock is still passed since we update its rate at
runtime depending on playback or capture sample rate.

Bug 200503387

Change-Id: I330a27eecbde99fdcb6359badfba773f20c3d1ec
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2159588
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>