tegra: nvdisp: disable cursor if offscreen
As per NvDisplay (T186/T194) hw manuals, if cursor is offscreen,
it is SW responsibility to disable cursor. Lack of this handling
was creating undefined cursor movement in extended desktop on dual
screen use-cases. Address it by disabling cursor if cursor is found
to be offscreen.

Note: similar issue remains on T21x but it is left as todo because
unlike NvDisplay, T21x clip window could be window A/B/C or Desktop.
This requires different calculations than NvDisplay and thus would be
addressed by Bug 2671921.

Bug 200420097
Bug 2608384

Change-Id: I4b48dbe3e4d137935becc26c0a125ff8b39038ad
Signed-off-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2164582
Reviewed-by: Shu Zhong <shuz@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Yogish Kulkarni <yogishk@nvidia.com>
Tested-by: Yogish Kulkarni <yogishk@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>