powergate: remove unnecessary WAR of audio
For t210, audio module need to be powered on at very early system boot,
but the powergate module is not ready at that time, so the WAR is that
we power on the audio module first and then set the unpowergate refcount
to "1" without unpowergating it during the init of powergate module. Due
to without unpowergating, there would be a mismatch, the counter maintained
by BPMP FW wouldn't indicated the real status of audio power domain.

Due to the update of AGIC, dont need this WAR any more, remove it.

Bug 2000015

Change-Id: I4b914f5f9b9e96696d273dee86bbb0fd440ea4c0
Signed-off-by: Taylor Xiao <txiao@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2153911
(cherry picked from commit a081d76866293045bc7fd0aa98794183b6a55a23)
Reviewed-on: https://git-master.nvidia.com/r/2169822
Reviewed-by: Phoenix Jung <pjung@nvidia.com>
Reviewed-by: Uday Gupta <udayg@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Timo Alho <talho@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>