eqos: fix clearing filter registers for MAC addresses
Issue: If any of the Multicast address deleted from filter
register driver is not resetting the corresponding address
in filter register which leads to allow the packets at driver
level which eventually forwarded to application.

Fix: Clear AE (address enable) bit for the address which is
removed from filtering list.

Bug 2657070

Change-Id: Iad93683e3b97381b86d0aecfb95609831df13d43
Signed-off-by: Bhadram Varka <vbhadram@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2166752
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>