video: tegra: host: Fix syncpt leak during detach
Both syncpoint channel attach and detach calls take a reference to
a syncpoint. During the attach call the reference is not dropped but
the reference becomes implicitly the channel reference. During
the detach call the code currently drops a single reference - thereby
leaving the originally taken reference intact.

This commit modifies the detach ioctl to drop two references (one that
was taken during the attach ioctl call and another that was taken
during the attach call).

Bug 200543776

Change-Id: I227588866046a810ee5d10144d0f26d3fb55ad82
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2176730
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-by: Deepak Nibade <dnibade@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>