t19x: common: add name-prefix for rt565x DAI link
The 'name-prefix' entry is removed from the base files where the default
state of audio DAI links are defined. The spdif-dit entries are being
removed in other patch in the current series and hence prefix is not
needed in the base files. Hence wherever the DT override is done for
DAI links, prefix is added as per the earlier entry.

Bug 200516191
Bug 200520821
Bug 200503387

Change-Id: Ic5d917fbaaa6c8a600602a6836a29b46bdc02955
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2177385
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>