t19x: galen: remove spdif-dit audio routing
The spdif-dit entries are being removed from the DAI links in other
patch in the series. The audio routing related to these is not more
needed and hence removed. The prefix name for rt565x audio codec is
moved to the corresponding link where other properties related to it
are overridden.

Bug 200516191
Bug 200520821
Bug 200503387

Change-Id: Icf90a7e45fc04ab43c81d3bf6cba93c31d1bc996
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2177386
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>