soc/tegra: fuse: Implement nvmem device
The nvmem framework provides a generic infrastructure and API to access
the type of information stored in fuses such as the Tegra FUSE block.

Implement an nvmem device that can be used to access the information in
a more generic way to decouple consumers from the custom Tegra API and
to add a more formal way of creating the dependency between the FUSE
device and the consumers.

Change-Id: Ib1ed00ba312dfd1384ef5d9d1566cfa464dc3c1a
Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit 96ee12b2a203167ffda7ac4a444418ca53df056d)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.4/+/2335843
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: Bitan Biswas <bbiswas@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit