t19x: galen: Add overlays for Jetson Xavier
Add the following device-tree overlays for the Jetson Xavier platform.

1. ADAFRUIT SPH0645LM4H (I2S MEMS Microphone)
2. FE-PI Audio Z V2 (SGTL5000 Audio Codec)
3. Jetson 40-pin Expansion Header

The Jetson 40-pin Expansion Header overlay has been developed for
the 'Jetson-IO' tool to help user re-configure the IOs exposed by
the header. This overlay adds a 'header-40pin-pinmux' node under
the pinmux device-node that defines the mapping of the SoC IOs to
the pins on the header. It also defines a default configuration for
the pins exposed by the header.

The overlays for the ADAFRUIT and FE-PI add-on modules are developed
to be directly applied to the base DTB file for the Jetson platform.
For example, these overlays can be used a follows in order to generate
a complete DTB with the overlay applied.

 $ fdtoverlay -i <base-dtb> -o <out-dtb> <overlay-dtbo>

Finally, device-tree include source files have been added to define
common phandles for items such as the 3.3V and 5V supplies, so that
overlays can use these phandles and work all the Jetson platforms.

Bug 2414131

Change-Id: If81cf7028d9ec72919ab5afa77959a2b352e608a
Signed-off-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2181413
(cherry picked from commit fac965e00d72946422403764a855456b0c4645c9)
Reviewed-on: https://git-master.nvidia.com/r/2217436
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>